% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_style.R
\name{dd_style}
\alias{dd_style}
\title{Add Dataloog theme to ggplot chart}
\usage{
dd_style()
}
\description{
This function allows you to add the Dataloog theme to your ggplotgraphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
dd_style()
}
\keyword{dd_style}
