% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_histogram.R
\name{emoji_histogram}
\alias{emoji_histogram}
\title{Create a histogram.}
\usage{
emoji_histogram(
  x = x,
  bins = 10,
  emoji = "🌸",
  emojiSize = 20,
  xticks = 10,
  yticks = 5,
  xtitle = "Value",
  ytitle = "Count",
  title = "Emoji Histogram",
  font = "Verdana, Geneva, Tahoma, sans-serif",
  xFontSize = 10,
  yFontSize = 10,
  xtitleFontSize = 14,
  ytitleFontSize = 14,
  titleFontSize = 18,
  stroke = "#444",
  strokeWidth = 0.5,
  width = 500,
  height = 300,
  opacity = 0.9,
  axisCol = "#555",
  bgcol = "#fffafc"
)
}
\arguments{
\item{x}{A vector of data.}

\item{bins}{The number of bins to consider. Defaults to 30.}

\item{emoji}{The emoji to use, defaults to "🌸"}

\item{xticks}{Optional. the number of x-axis ticks to consider.}

\item{yticks}{Optional. The number of y-axis ticks to consider.}

\item{xtitle}{Optional. The title of the x-axis.}

\item{ytitle}{Optional. The title of the y-axis.}

\item{title}{Optional. The title of the plot.}

\item{font}{The font family to consider for the titles. Defaults to
"Verdana, Geneva, Tahoma, sans-serif".}

\item{xFontSize}{the font size of the x-axis labels. Defaults to 10.}

\item{yFontSize}{the font size of the y-axis labels. Defaults to 10.}

\item{xtitleFontSize}{The font size of the x-axis title. Defaults to 16.}

\item{ytitleFontSize}{The font size of the y-axis title. Defaults to 16.}

\item{titleFontSize}{The font size of the plot title. Defaults to 22.}

\item{stroke}{The stroke color of the bars. Defaults to 'crimson'.}

\item{strokeWidth}{Optional. the stroke width of the bars.}

\item{width}{Optional. The width of the SVG output.}

\item{height}{Optional. The height of the SVG output.}

\item{opacity}{The color opacity of the bars (from 0 to 1). Defaults to 1.}

\item{axisCol}{the color of the x and y axis. It includes the ticks, the labels and titles.
Defaults to 'black'.}

\item{bgcol}{The background color of the SVG. Defaults to "#CAD0D3" HEX color.}

\item{emoji_size}{The size of the emoji. Defaults to 20}

\item{fill}{The color of the bars. Defaults to 'crimson'.}
}
\value{
A SVG histogram.
}
\description{
Create a histogram.
}
\examples{

histogram(
  x = mtcars$mpg,
  emoji = "🌸",
  bins = 20,
  fill = "crimson",
  stroke = "white",
  strokeWidth = 1,
  title = "Distribution of the hwy variable",
  width = "20",
  height = "10"
)
}
