% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dereck.R
\name{theme_dereck}
\alias{theme_dereck}
\title{Add derecksnotes theme to ggplot chart}
\usage{
theme_dereck(base_size = 11, base_family = "")
}
\description{
This function allows you to add the derecksnotes theme to your ggplotgraphics.
}
\examples{
iris \%>\%
ggplot(aes(x = Sepal.Length, y = Sepal.Width)) +
geom_point(aes(color = Species, shape = Species)) +
labs(title = "Iris data", subtitle = "Classic data", x = "Sepal Length", y = "Sepal Width") +
facet_grid(~ Species) +
theme_dereck()

data.frame(year = time(nhtemp), tmps = as.matrix(nhtemp)) \%>\%
ggplot(aes(x = year, y = tmps)) +
geom_line() +
# geom_hline(yintercept = 0, size = 1, colour = "#333333") +
labs(title = "New Hampshire yearly temperatures", subtitle = "Collected between 1912 and 1971", x = "Year", y = "Average temperature") +
theme_dereck()

}
\keyword{theme_derecksnotes}
