% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Create a scatter plot}
\usage{
scatterPlot(
  data,
  x,
  y,
  col = "crimson",
  size = 2,
  id = NULL,
  xticks = NULL,
  yticks = NULL,
  xtitle = NULL,
  xtitleFontSize = 16,
  ytitle = NULL,
  ytitleFontSize = 16,
  title = NULL,
  titleFontSize = 22,
  stroke = NULL,
  strokeWidth = NULL,
  font = "Verdana, Geneva, Tahoma, sans-serif",
  bgcol = "#CAD0D3",
  opacity = 1,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{the data frame containing the quantitative variables}

\item{x}{the x-variable to plot.}

\item{y}{the y-variable to plot}

\item{col}{the color of the dots. Defaults to 'crimson'}

\item{size}{the size of the dots. Defaults to 2}

\item{id}{optional. the id of the SVG output.}

\item{xticks}{optional. the number of x-axis ticks to consider}

\item{yticks}{optional. the number of y-axis ticks to consider}

\item{xtitle}{optional. the title of the x-axis}

\item{xtitleFontSize}{the font size of the x-axis title. Defaults to 16}

\item{ytitle}{optional. the title of the y-axis}

\item{ytitleFontSize}{the font size of the y-axis title. Defaults to 16}

\item{title}{optional. the title of the plot}

\item{titleFontSize}{the font size of the plot title. Defaults to 22}

\item{stroke}{optional. the stroke color of the dots}

\item{strokeWidth}{optional. the stroke width of the dots}

\item{font}{the font family to consider for the titles. Defaults to
"Verdana, Geneva, Tahoma, sans-serif"}

\item{bgcol}{the background color of the SVG. Defaults to "#CAD0D3" HEX color}

\item{opacity}{the color opacity of the dots (from 0 to 1). Defaults to 1}

\item{width}{optional. the width of the SVG output}

\item{height}{optional. the height of the SVG output}
}
\value{
a SVG scatter plot
}
\description{
Create a scatter plot
}
\examples{

scatterPlot(
  data = mpg,
  x = "cty",
  y = "hwy"
  )



}
